+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 CHANGELOG:
 
+# 1.4.0
+
+Dependency changed: LitJson (new).
+
+New features:
+- Cache remote config data as JSON to PlayerPref.
+- Add default config as scriptable object.
+- Make new extra package that include Firebase manager prefab, Default config files, Firebase cloud messaging helper.
+- Accessing Firebase manager instance will automatically instantiate an instance using Resouces.Load.
+
 # 1.3.1
 Changes:
 - Clean up code.